Misc improvements
