+++ .github/workflows/fetch_contributors.yml
@@ -4,8 +4,6 @@ on:
   pull_request: # This will run the workflow on pull request events
     types:
       - closed
-  schedule: # This will run the workflow on a schedule
-    - cron: '0 0 * * 0'
   workflow_dispatch: # This will run the workflow on manual trigger
 
 permissions:
@@ -13,34 +11,34 @@ permissions:
 
 jobs:
   contrib-readme-job:
-    name: A job to automate contrib in readme
+    name: A job to fetch contributors list
     runs-on: ubuntu-latest
 
     steps:
       - name: Checkout
         uses: actions/checkout@v4
 
-      - name: Contribute List
+      - name: Fetch contributors list
         uses: akhilmhdh/contributors-readme-action@v2.3.10
         with:
-          image_size: 32
-          readme_path: "docs/contributors.md"
+          image_size: 100
+          readme_path: ./docs/contributors.md
           collaborators: all
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-      - name: Commit changes
+      - name: Save contributors in assets
         if: github.event_name == 'workflow_dispatch' || github.event.pull_request.merged == true
         run: |
-          printf '%s\n' $(curl "https://api.github.com/repos/DHD2280/Oxygen-Customizer/contributors") > ./app/src/main/assets/Misc/contributors.json
+          printf '%s\n' $(curl "https://api.github.com/repos/Mahmud0808/Iconify/contributors") > ./app/src/main/assets/Misc/contributors.json
 
-      - name: Commit
+      - name: Commit changes and push
         uses: github-actions-x/commit@v2.9
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           push-branch: 'beta'
-          commit-message: 'Fetch Contributors'
+          commit-message: 'Update contributors list'
           force-add: 'true'
           files: app/src/main/assets/Misc/contributors.json docs/contributors.md
-          name: GitHub
-          email: OxygenCustomizer@dhd.it
+          name: Mahmud0808
+          email: crazymahmud08@gmail.com
